pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vacant-lot-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "earthengine-api==1.7.1",
    "fiona>=1.10.1",
    "geemap>=0.36.4",
    "geopandas>=1.1.1",
    "google-api-python-client==2.187.0",
    "google-auth>=2.41.1",
    "google-cloud-storage>=3.4.1",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "nbformat>=5.10.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.0",
    "pyyaml>=6.0.3",
    "rasterio>=1.5.0",
    "scikit-image>=0.21",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "pystac-client>=0.7.0",
    "planetary-computer>=1.0.0",
    "requests>=2.32.5",
    "nbstripout>=0.8.1",
    "pygris>=0.2.1",
    "torch>=2.0",
    "torchvision>=0.15",
    "kaggle>=1.6",
    "lightgbm>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["vacant_lot"]
